requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
openai==1.12.0
Pillow==10.2.0
pydantic==2.6.1
pydantic-core==2.16.2
typing-extensions>=4.9.0